# Conflicts:
#	src/main/java/fr/eni/lesobjetssontnosamis/bll/ArticleVenduServiceImpl.java
#	src/main/java/fr/eni/lesobjetssontnosamis/bll/UtilisateurServiceImpl.java
#	src/main/java/fr/eni/lesobjetssontnosamis/controller/UtilisateurController.java
#	src/main/java/fr/eni/lesobjetssontnosamis/dal/ArticleVenduDAOImpl.java
#	src/main/java/fr/eni/lesobjetssontnosamis/dal/UtilisateurDAOImpl.java
#	src/main/resources/templates/fragments/frag-header.html

+++ src/main/java/fr/eni/lesobjetssontnosamis/bll/ArticleVenduService.java
@@ -12,7 +12,7 @@ public interface ArticleVenduService {
 
     List<ArticleVendu> getArticleVendus();
 
-    ArticleVendu findArticleVendu(int noArticle);
+    ArticleVendu findArticleVendu(long noArticle);
 
     void addArticleVendu(ArticleVendu articleVendu);
 }
