+++ src/main/java/fr/eni/lesobjetssontnosamis/bo/Categorie.java
@@ -8,7 +8,7 @@ public class Categorie {
     // ATRIBUTS
     //--------------------------------
 
-    private int noCategorie;
+    private long noCategorie;
     private String libelle;
 
     // Une catégorie par article et plusieurs articles pour chaque catégorie
@@ -37,10 +37,10 @@ public class Categorie {
     // GETTERS & SETTERS
     //--------------------------------
 
-    public int getNoCategorie() {
+    public long getNoCategorie() {
         return noCategorie;
     }
-    public void setNoCategorie(int noCategorie) {
+    public void setNoCategorie(long noCategorie) {
         this.noCategorie = noCategorie;
     }
