+++ src/main/java/fr/eni/lesobjetssontnosamis/bo/Utilisateur.java
@@ -17,7 +17,7 @@ public class Utilisateur implements UserDetails {
     // ATTRIBUTS
     //--------------------------------
     private static final long serialVersionUID = 1L;
-    private int noUtilisateur;
+    private long noUtilisateur;
 
     // @NotBlank
     // @Size(min = 2, max=250, message = "Pseudo non valide")
@@ -70,7 +70,7 @@ public class Utilisateur implements UserDetails {
     }
 
     public Utilisateur(int noUtilisateur, String pseudo, String nom, String prenom, String email, String telephone, String rue, int codePostal, String ville, String motDePasse, int credit, boolean administrateur) {
-        this(); // appel du constructeur par défaut
+        this();
         this.noUtilisateur = noUtilisateur;
         this.pseudo = pseudo;
         this.nom = nom;
@@ -89,10 +89,10 @@ public class Utilisateur implements UserDetails {
     // GETTERS & SETTERS
     // --------------------------------
 
-    public int getNoUtilisateur() {
+    public long getNoUtilisateur() {
         return noUtilisateur;
     }
-    public void setNoUtilisateur(int noUtilisateur) {
+    public void setNoUtilisateur(long noUtilisateur) {
         this.noUtilisateur = noUtilisateur;
     }
